Yet more minor linting refactors.
