+++ src/utils/convertToMB.ts
@@ -3,6 +3,8 @@
  *  Licensed under the MIT License. See LICENSE.md in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+/* eslint-disable unicorn/filename-case */
+
 export function convertToMB(bytes: number): number {
     return Math.round(bytes / 1000000);
 }
