+++ src/utils/wrapDockerodeENOENT.ts
@@ -3,6 +3,8 @@
  *  Licensed under the MIT License. See LICENSE.md in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+/* eslint-disable unicorn/filename-case */
+
 import { parseError } from "vscode-azureextensionui";
 
 export async function wrapDockerodeENOENT<T>(dockerodeCallback: () => Promise<T>): Promise<T> {
