--- tslint.json
@@ -1,137 +1,0 @@
-{
-    "extends": "tslint-microsoft-contrib",
-    "rules": {
-        "no-reserved-keywords": false,
-        "await-promise": [
-            true,
-            "Thenable",
-            "RequestPromise"
-        ],
-        "no-console": [
-            false,
-            "debug",
-            "info",
-            "log",
-            "time",
-            "timeEnd",
-            "trace"
-        ],
-        "no-empty": false,
-        "no-increment-decrement": false,
-        "no-use-before-declare": false, // (recommended turned off: https://palantir.github.io/tslint/rules/no-use-before-declare/)
-        "restrict-plus-operands": false,
-        "strict-boolean-expressions": [
-            false,
-            "allow-string",
-            "allow-undefined-union",
-            "allow-mix"
-        ],
-        "completed-docs": [
-            false,
-            "classes"
-        ],
-        "export-name": true,
-        "function-name": false,
-        "import-name": false,
-        "interface-name": false,
-        "max-classes-per-file": [
-            false,
-            3
-        ],
-        "max-line-length": [
-            false,
-            140
-        ],
-        "member-ordering": [
-            false,
-            {
-                "order": "fields-first"
-            }
-        ],
-        "no-default-export": false,
-        "no-null-keyword": false,
-        "no-parameter-properties": false,
-        "no-relative-imports": false,
-        "no-require-imports": false,
-        "no-suspicious-comment": false, // TODO
-        "no-unnecessary-local-variable": false,
-        "no-void-expression": [
-            false,
-            "ignore-arrow-function-shorthand"
-        ],
-        "one-variable-per-declaration": false,
-        "prefer-const": false,
-        "prefer-template": false,
-        "typedef": [
-            true,
-            "call-signature",
-            // "arrow-call-signature", changed
-            "parameter",
-            // "arrow-parameter", changed
-            "property-declaration",
-            // "variable-declaration", changed
-            "member-variable-declaration"
-        ],
-        "variable-name": [
-            true,
-            "allow-leading-underscore",
-            "ban-keywords"
-        ],
-        "eofline": false,
-        "linebreak-style": false,
-        "newline-before-return": false,
-        "no-single-line-block-comment": false,
-        "object-literal-key-quotes": [
-            false,
-            "as-needed"
-        ],
-        "quotemark": [
-            false,
-            "single"
-        ],
-        "semicolon": [
-            false,
-            "always"
-        ],
-        "trailing-comma": [
-            false,
-            {
-                "singleline": "never",
-                "multiline": "never"
-            }
-        ],
-        "no-unexternalized-strings": [
-            false,
-            {
-                "signatures": [
-                    "localize",
-                    "nls.localize"
-                ],
-                "keyIndex": 0,
-                "messageIndex": 1
-            }
-        ],
-        "prefer-type-cast": false,
-        "no-implicit-dependencies": [
-            true,
-            [
-                "vscode"
-            ]
-        ],
-        "newline-per-chained-call": false,
-        "no-submodule-imports": false,
-        "type-literal-delimiter": false,
-        "prefer-readonly": false,
-        "no-return-await": false,
-        "no-parameter-reassignment": false,
-        "no-object-literal-type-assertion": false,
-        "no-duplicate-imports": false,
-        "no-backbone-get-set-outside-model": false,
-        "binary-expression-operand-order": false
-    },
-    "linterOptions": {
-        "exclude": [
-            "test/**"
-        ]
-    }
-}
